SCHED-75: Moving tests to test directory and moving outputter to common package.
