+++ common/output/__init__.py
@@ -3,9 +3,10 @@ import json
 from typing import NoReturn
 
 from api.ocs import OcsProgramProvider
-from common.minimodel import Group, Observation
+from common.minimodel import Group, Observation, Program
 
-if __name__ == '__main__':
+
+def print_ocs_program(filename=os.path.join('data', 'GN-2018B-Q-101.json')) -> NoReturn:
     provider = OcsProgramProvider
     # data = provider.load_program(os.path.join('data', 'GN-2018B-Q-101.json'))
     path = os.path.join('data', 'GN-2018B-Q-101.json')
@@ -13,6 +14,10 @@ if __name__ == '__main__':
         data = json.loads(f.read())
 
     program = OcsProgramProvider.parse_program(data['PROGRAM_BASIC'])
+    print_program(program)
+
+
+def print_program(program: Program) -> NoReturn:
     print(f'Program: {program.id}')
 
     def sep(depth: int) -> str:
